DisplayCommandHandler
Reference:https://github.com/xen-troops/displ_be/issues/118

Just check for the nullptr is added.
The reason is, because the parameters to create the connector are comming
in some ways, it means, it is necessary to gather them and save somewher to create the
connector when all required parameters are ready. It does not bring any advantage.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>